New utils.chunk dir. Make chunk_loader lower case.
