+++ napari/utils/chunk/__init__.py
@@ -1,0 +1,6 @@
+"""Synchronous and Asynchronous Chunk Loading.
+"""
+import os
+
+from ._request import ChunkRequest
+from ._loader import chunk_loader, synchronous_loading
